+++ functions/src/index.ts
@@ -13,9 +13,9 @@ export { onVoteRequestCreate } from "./definitions/eventHandlers/onVoteRequestCr
 export { onVoteRequestUpdateV2 } from "./definitions/eventHandlers/onVoteRequestUpdate"
 export { healthcheckV2 } from "./definitions/healthcheck"
 export { onMessageWriteV2 } from "./definitions/eventHandlers/updateTypesense"
-export { onUserPublish } from "./definitions/eventHandlers/userHandlers"
+export { onUserMessagePublish } from "./definitions/eventHandlers/userHandlers"
 export { onCheckerPublish } from "./definitions/eventHandlers/checkerHandlerWhatsapp"
-export { onUserInteractivePublish } from "./definitions/eventHandlers/userHandlerWhatsappInteractive"
+export { onUserInteractivePublish } from "./definitions/eventHandlers/userInteractionHandlers"
 export { telegramAuthHandler } from "./definitions/api/authentication"
 export { onCheckerPublishTelegram } from "./definitions/eventHandlers/checkerHandlerTelegram"
 export { apiHandler } from "./definitions/api/api"
